requirements.txt
accelerate==1.8.1
biopython==1.85
build==1.3.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
cryptography==45.0.4
cycler==0.12.1
fair-esm==2.0.0
faiss-cpu==1.11.0.post1
filelock==3.18.0
fonttools==4.58.4
fsspec==2025.5.1
hf-xet==1.1.5
huggingface-hub==0.33.2
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.5
numpy==2.3.1
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
opencv-python==4.11.0.86
packaging==25.0
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.2.1
pip-tools==7.5.0
polars==1.34.0
polars-runtime-32==1.34.0
psutil==7.0.0
pycparser==2.22
pyfastx==2.2.0
PyMuPDF==1.26.1
pyparsing==3.2.3
PyPDF2==3.0.1
pypdfium2==4.30.1
pyproject_hooks==1.2.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.2
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.53.1
triton==3.3.1
typing_extensions==4.14.1
urllib3==2.5.0
